+++ src/entities/enums/gender.enum.ts
@@ -1,0 +1,4 @@
+export enum Gender {
+  MALE = 'M',
+  FEMALE = 'F'
+}
