+++ src/entities/enums/question-type.enum.ts
@@ -1,0 +1,4 @@
+export enum QuestionType {
+  MULTIPLE_CHOICES = 'M',
+  SINGLE_CHOICE = 'S'
+}
