+++ src/entities/enums/role.enum.ts
@@ -1,0 +1,5 @@
+export enum Role {
+  ADMIN = 'admin',
+  TEACHER = 'teacher',
+  STUDENT = 'student'
+}
